cp3-ajuste

+++ src/services/firebaseActions.tsx
@@ -1,12 +1,10 @@
 import { collection, addDoc, deleteDoc, doc, getDocs, updateDoc, query, where } from "firebase/firestore";
 import { db } from '../services/firebaseConfig';
 
-
-
-// Função para adicionar filme aos favoritos com ranking
-export const addFavoriteMovie = async (userId, movie, ranking) => {
+// Função para adicionar filme aos favoritos sem ranking
+export const addFavoriteMovie = async (userId, movie) => {
   try {
-    const docRef = await addDoc(collection(db, 'favoriteMovies'), { ...movie, userId, ranking });
+    const docRef = await addDoc(collection(db, 'favoriteMovies'), { ...movie, userId });
     console.log("Filme adicionado aos favoritos com ID: ", docRef.id);
   } catch (e) {
     console.error("Erro ao adicionar filme aos favoritos: ", e);
@@ -15,6 +13,11 @@ export const addFavoriteMovie = async (userId, movie, ranking) => {
 
 // Função para remover filme dos favoritos
 export const removeFavoriteMovie = async (docId) => {
+  if (!docId) {
+    console.error("docId está indefinido");
+    return;
+  }
+
   try {
     await deleteDoc(doc(db, 'favoriteMovies', docId));
     console.log("Filme removido dos favoritos com ID: ", docId);
@@ -25,14 +28,29 @@ export const removeFavoriteMovie = async (docId) => {
 
 // Função para listar filmes favoritos de um usuário
 export const getFavoriteMovies = async (userId) => {
-  const q = query(collection(db, 'favoriteMovies'), where("userId", "==", userId));
-  const querySnapshot = await getDocs(q);
-  const favoriteMovies = querySnapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
-  return favoriteMovies;
+  if (!userId) {
+    console.error("userId está indefinido");
+    return [];
+  }
+
+  try {
+    const q = query(collection(db, 'favoriteMovies'), where("userId", "==", userId));
+    const querySnapshot = await getDocs(q);
+    const favoriteMovies = querySnapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
+    return favoriteMovies;
+  } catch (e) {
+    console.error("Erro ao carregar filmes favoritos: ", e);
+    return [];
+  }
 };
 
-// Função para atualizar dados de um filme favorito, incluindo ranking
+// Função para atualizar dados de um filme favorito
 export const updateFavoriteMovie = async (docId, updatedData) => {
+  if (!docId || !updatedData) {
+    console.error("docId ou updatedData está indefinido");
+    return;
+  }
+
   try {
     const docRef = doc(db, 'favoriteMovies', docId);
     await updateDoc(docRef, updatedData);
